Move number.ts to /common
