+++ extensions/ql-vscode/src/common/number.ts
