+++ extensions/ql-vscode/src/view/variant-analysis/VariantAnalysisRepositoriesStats.tsx
@@ -1,6 +1,6 @@
 import * as React from "react";
 import { VariantAnalysisStatus } from "../../variant-analysis/shared/variant-analysis";
-import { formatDecimal } from "../../pure/number";
+import { formatDecimal } from "../../common/number";
 import {
   ErrorIcon,
   HorizontalSpace,
