+++ extensions/ql-vscode/test/unit-tests/pure/number.test.ts
@@ -1,4 +1,4 @@
-import { formatDecimal } from "../../../src/pure/number";
+import { formatDecimal } from "../../../src/common/number";
 
 describe("Number", () => {
   it("should return a formatted decimal", () => {
